requirements.txt
matplotlib
jupyter==1.0.0
nvidia-dali-cuda110==0.26.0
--extra-index-url https://developer.download.nvidia.com/compute/redist

